requirements.txt
tensorflow
tensorflow-addons
sklearn
tqdm
pandas
biopython
snakemake
pyyaml
complexcgr==0.7.4